[refactor] all the classes earlier in odemis.gui.comp.overlay.world.py have been split into individual python files.
